deleted reply db thats being created by another function

+++ full-project/backend/app/repos/replyRepo.py
@@ -11,7 +11,11 @@ def loadAll():
         return json.load(f)
     
 def saveAll(replies):
-    #save all replies to the JSON file
-    os.makedirs(os.path.dirname(DATA_PATH), exist_ok=True)
+    if not os.path.exists(os.path.dirname(DATA_PATH)):
+        
+        return False
+
     with open(DATA_PATH, "w", encoding="utf-8") as f:
-        json.dump(replies, f, indent=2,)
+        json.dump(replies, f, indent=2, ensure_ascii=False)
+
+    return True
